(PC-34411) feat(chronicle): avoid code duplication
