+++ src/features/chronicle/pages/Chronicles/Chronicles.web.tsx
@@ -1,6 +1,5 @@
-import { useNavigation, useRoute } from '@react-navigation/native'
-import React, { FunctionComponent, useCallback, useRef } from 'react'
-import { FlatList, InteractionManager } from 'react-native'
+import { useRoute } from '@react-navigation/native'
+import React, { FunctionComponent } from 'react'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import styled, { useTheme } from 'styled-components/native'
 
@@ -8,12 +7,10 @@ import { SubcategoryIdEnumv2 } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { offerChroniclesToChronicleCardData } from 'features/chronicle/adapters/offerChroniclesToChronicleCardData/offerChroniclesToChronicleCardData'
 import { useChronicles } from 'features/chronicle/api/useChronicles/useChronicles'
-import { ChronicleCardList } from 'features/chronicle/components/ChronicleCardList/ChronicleCardList'
 import { ChronicleOfferInfo } from 'features/chronicle/components/ChronicleOfferInfo/ChronicleOfferInfo.web'
-import { ChroniclesHeader } from 'features/chronicle/components/ChroniclesHeader/ChroniclesHeader'
-import { ChroniclesWebMetaHeader } from 'features/chronicle/components/ChroniclesWebMetaHeader/ChroniclesWebMetaHeader'
+import { ChroniclesBase } from 'features/chronicle/pages/Chronicles/ChroniclesBase'
 import { ChronicleCardData } from 'features/chronicle/type'
-import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
+import { UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useOffer } from 'features/offer/api/useOffer'
 import { OfferCTAButton } from 'features/offer/components/OfferCTAButton/OfferCTAButton'
 import { getOfferPrices } from 'features/offer/helpers/getOfferPrice/getOfferPrice'
@@ -22,14 +19,10 @@ import { getDisplayedPrice } from 'libs/parsers/getDisplayedPrice'
 import { useSubcategoriesMapping } from 'libs/subcategories'
 import { useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { useGetPacificFrancToEuroRate } from 'shared/exchangeRates/useGetPacificFrancToEuroRate'
-import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
-import { TypoDS, getSpacing } from 'ui/theme'
 
 export const Chronicles: FunctionComponent = () => {
   const route = useRoute<UseRouteType<'Chronicles'>>()
   const offerId = route.params?.offerId
-  const chronicleId = route.params?.chronicleId
-  const { navigate } = useNavigation<UseNavigationType>()
   const { data: offer } = useOffer({ offerId })
   const subcategoriesMapping = useSubcategoriesMapping()
 
@@ -46,13 +39,10 @@ export const Chronicles: FunctionComponent = () => {
     ? subcategoriesMapping[offer.subcategoryId]
     : subcategoriesMapping[SubcategoryIdEnumv2.CONCERT]
 
-  const { headerTransition, onScroll } = useOpacityTransition()
   const { trackEventHasSeenOfferOnce } = useOfferBatchTracking(subcategory.id)
   const prices = getOfferPrices(offer?.stocks ?? [])
   const currency = useGetCurrencyToDisplay()
   const euroToPacificFrancRate = useGetPacificFrancToEuroRate()
-
-  const chroniclesListRef = useRef<FlatList<ChronicleCardData>>(null)
 
   const displayedPrice = getDisplayedPrice(
     prices,
@@ -62,50 +52,15 @@ export const Chronicles: FunctionComponent = () => {
     { fractionDigits: 2 }
   )
 
-  const selectedChronicle = chronicleCardsData?.findIndex((item) => item.id === chronicleId) ?? -1
-
-  const handleLayout = useCallback(() => {
-    if (selectedChronicle !== -1) {
-      InteractionManager.runAfterInteractions(() => {
-        chroniclesListRef.current?.scrollToIndex({
-          index: selectedChronicle,
-          animated: true,
-          viewOffset: headerHeight,
-        })
-      })
-    }
-  }, [selectedChronicle, headerHeight])
-
-  const handleGoBack = () => {
-    navigate('Offer', { id: offerId, openModalOnNavigation: undefined })
-  }
-
   if (!offer || !chronicleCardsData) return null
-
-  const title = `Tous les avis sur "${offer.name}"`
-
-  const listComponent = (
-    <StyledChronicleCardList
-      ref={chroniclesListRef}
-      data={chronicleCardsData}
-      onScroll={onScroll}
-      paddingTop={headerHeight}
-      headerComponent={<StyledTitle2>Tous les avis</StyledTitle2>}
-      onLayout={handleLayout}
-    />
-  )
 
   return (
     <Container>
-      <ChroniclesWebMetaHeader title={title} />
-      <ChroniclesHeader
-        handleGoBack={handleGoBack}
-        headerTransition={headerTransition}
-        title={title}
-      />
-
-      {isDesktopViewport ? (
-        <FullFlexRow>
+      <ChroniclesBase
+        offerId={offer.id}
+        offerName={offer.name}
+        chronicleCardsData={chronicleCardsData}>
+        {isDesktopViewport ? (
           <StyledChronicleOfferInfo
             imageUrl={offer.images?.[0]?.url ?? ''}
             title={offer.name}
@@ -118,46 +73,27 @@ export const Chronicles: FunctionComponent = () => {
               trackEventHasSeenOfferOnce={trackEventHasSeenOfferOnce}
             />
           </StyledChronicleOfferInfo>
-          {listComponent}
-        </FullFlexRow>
-      ) : (
-        listComponent
-      )}
+        ) : null}
+      </ChroniclesBase>
     </Container>
   )
 }
-
-const StyledTitle2 = styled(TypoDS.Title2)({
-  marginBottom: getSpacing(6),
-})
 
 const FullFlexView = styled.View({
   flex: 1,
 })
 
-const FullFlexRow = styled(FullFlexView)({
-  flexDirection: 'row',
-  columnGap: getSpacing(18),
-})
-
 const Container = styled(FullFlexView)(({ theme }) => ({
-  paddingHorizontal: theme.contentPage.marginHorizontal,
-  paddingTop: theme.contentPage.marginVertical,
+  ...(theme.isDesktopViewport
+    ? {
+        paddingHorizontal: theme.contentPage.marginHorizontal,
+        paddingTop: theme.contentPage.marginVertical,
+      }
+    : {}),
 }))
-
-const StyledChronicleCardList = styled(ChronicleCardList).attrs<{ paddingTop: number }>(
-  ({ paddingTop }) => ({
-    horizontal: false,
-    separatorSize: 6,
-    contentContainerStyle: { paddingVertical: paddingTop },
-  })
-)<{ paddingTop: number }>({
-  flex: 1,
-})
 
 const StyledChronicleOfferInfo = styled(ChronicleOfferInfo)<{ paddingTop: number }>(
   ({ paddingTop }) => ({
     paddingTop,
-    flex: 1,
   })
 )
